Address review: server searchParams + SEO + cruft cleanup
Fixes from the multizone and zone reviews:

- **UK routing** was silently rendering US because Next.js rewrites are
  reverse proxies — the ?country=uk on the rewrite destination reaches
  the zone's Node runtime but never the browser's location.search. Split
  page.jsx into a server component that reads searchParams.country and
  passes initialCountry as a prop to a new client component, MarriageApp.
  Seeded useState with the prop so /uk/marriage hydrates straight into UK
  without a flash of the US default.
- **Stale GitHub Pages deploy workflow** (.github/workflows/deploy.yml)
  deleted — it pushed a Vite dist via npm, both of which are gone after
  the port. Vercel handles deploys.
- **SEO URLs** pointed at the dead policyengine.github.io site. Updated
  public/robots.txt and public/sitemap.xml to the first-party canonical
  https://policyengine.org/us/marriage, and deleted the GitHub Pages
  404.html redirect shim (Next has its own _not-found page).
- **README** rewritten — was still describing the original Streamlit app.
- Dropped dead CSS classes (.metric-cards-section, .sf-nyc-toggle) and
  Vite entries from .gitignore (dist/, .vite/).
