--- public/404.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Marriage Tax Calculator | PolicyEngine</title>
-  <script>
-    // SPA redirect: GitHub Pages serves 404.html for unknown paths.
-    // Redirect to index with the path preserved as a hash fragment.
-    var path = window.location.pathname.replace('/marriage/', '');
-    window.location.replace('/marriage/' + '#' + path);
-  </script>
-</head>
-<body></body>
-</html>
