+++ public/sitemap.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
-    <loc>https://policyengine.github.io/marriage/</loc>
-    <lastmod>2026-02-16</lastmod>
+    <loc>https://policyengine.org/us/marriage</loc>
+    <lastmod>2026-04-16</lastmod>
     <changefreq>monthly</changefreq>
   </url>
 </urlset>
